+++ examples/demo.zig
@@ -43,10 +43,15 @@ pub fn main() !void {
     const wall_diff = try this_time_Paris.diffWall(this_time_LA);
     const abs_diff = this_time_Paris.diff(this_time_LA);
 
-    std.debug.print(
-        "Wall clock time difference: {s}\nAbsolute time difference: {s}\n",
-        .{ wall_diff, abs_diff },
-    );
+    std.debug.print("Wall clock time difference: {s}\nAbsolute time difference: {s}\n\n", .{ wall_diff, abs_diff });
     // Wall clock time difference: PT9H
     // Absolute time difference: PT0S
+
+    // Easteregg:
+    const now = zdt.Datetime.nowUTC();
+    const easter_date = try zdt.Datetime.EasterDate(now.year);
+    buf.clearAndFree();
+    try easter_date.toString("%B %d, %Y", buf.writer());
+    std.debug.print("Easter this year is on {s}\n", .{buf.items});
+    // Easter this year is on April 20, 2025
 }